Move SlottedPipeBuilderTest to slotted-runtime module
